- Removed prismjs -  updated singview to use aceditor
