+++ client/package.json
@@ -10,7 +10,6 @@
     "@stripe/stripe-js": "^1.3.1",
     "ace-builds": "^1.4.9",
     "axios": "^0.19.2",
-    "prismjs": "^1.20.0",
     "react": "^16.8.6",
     "react-ace": "^8.1.0",
     "react-dom": "^16.8.6",
